Casting every error code to i32

+++ src/core/error.h
@@ -165,7 +165,7 @@ typedef struct {
 #define error_msg() (__error_context__->info.msg)
 #define error_code(T) ((T)__error_context__->info.code)
 
-#define error_set_code(code) _error_set_code(__error_context__, code)
+#define error_set_code(code) _error_set_code(__error_context__, (i32)code)
 #define error_set_msg(...) _error_set_msg(__error_context__, __VA_ARGS__)
 
 #define error_add_location(...) _error_add_location(__error_context__, FILE_LOC)
